Refactoring kafka source into seprate module, add more unit tests
